Added tslint and codelyser
