+++ app/app.component.ts
@@ -2,7 +2,7 @@ import { Component } from "@angular/core";
 import { NS_ROUTER_DIRECTIVES } from "nativescript-angular/router";
 
 @Component({
-  selector: "main",
+  selector: "gr-main",
   directives: [NS_ROUTER_DIRECTIVES],
   template: "<page-router-outlet></page-router-outlet>"
 })
