+++ app/auth.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Router, CanActivate } from '@angular/router';
+import { Router, CanActivate } from "@angular/router";
 import { LoginService } from "./shared";
 
 @Injectable()
