+++ app/shared/status-bar-util.ts
@@ -41,6 +41,6 @@ export function setStatusBarColors() {
           | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
           | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY);
       }
-    }
+    };
   }
 }
