+++ package.json
@@ -1,56 +1,61 @@
 {
-	"name": "Groceries",
-	"version": "1.0.0",
-	"description": "A NativeScript-built iOS and Android app for managing grocery lists",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/nativescript/sample-Groceries.git"
-	},
-	"keywords": [
-		"NativeScript"
-	],
-	"author": "TJ VanToll <tj.vantoll@gmail.com> (https://www.tjvantoll.com/)",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/nativescript/sample-Groceries/issues"
-	},
-	"homepage": "https://github.com/nativescript/sample-Groceries/groceries",
-	"nativescript": {
-		"id": "org.nativescript.groceries",
-		"tns-ios": {
-			"version": "2.2.0"
-		},
-		"tns-android": {
-			"version": "2.2.0"
-		}
-	},
-	"dependencies": {
-		"@angular/common": "2.0.0-rc.4",
-		"@angular/compiler": "2.0.0-rc.4",
-		"@angular/core": "2.0.0-rc.4",
-		"@angular/platform-browser": "2.0.0-rc.4",
-		"@angular/platform-browser-dynamic": "2.0.0-rc.4",
-		"@angular/platform-server": "2.0.0-rc.4",
-		"@angular/router": "3.0.0-beta.2",
-		"email-validator": "1.0.4",
-		"everlive-sdk": "1.7.1",
-		"nativescript-angular": "0.3.0",
-		"nativescript-iqkeyboardmanager": "1.0.1",
-		"nativescript-social-share": "1.3.0",
-		"nativescript-unit-test-runner": "0.3.3",
-		"tns-core-modules": "2.2.0"
-	},
-	"devDependencies": {
-		"babel-traverse": "6.8.0",
-		"babel-types": "6.8.1",
-		"babylon": "6.8.0",
-		"filewalker": "0.1.2",
-		"jasmine-core": "^2.4.1",
-		"karma": "^0.13.22",
-		"karma-jasmine": "^1.0.2",
-		"karma-nativescript-launcher": "^0.4.0",
-		"lazy": "1.0.11",
-		"nativescript-dev-typescript": "^0.3.2",
-		"typescript": "^1.8.10"
-	}
+  "name": "Groceries",
+  "version": "1.0.0",
+  "description": "A NativeScript-built iOS and Android app for managing grocery lists",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/nativescript/sample-Groceries.git"
+  },
+  "keywords": [
+    "NativeScript"
+  ],
+  "author": "TJ VanToll <tj.vantoll@gmail.com> (https://www.tjvantoll.com/)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/nativescript/sample-Groceries/issues"
+  },
+  "homepage": "https://github.com/nativescript/sample-Groceries/groceries",
+  "nativescript": {
+    "id": "org.nativescript.groceries",
+    "tns-ios": {
+      "version": "2.2.0"
+    },
+    "tns-android": {
+      "version": "2.2.0"
+    }
+  },
+  "scripts": {
+    "tslint": "tslint \"app/**/*.ts\""
+  },
+  "dependencies": {
+    "@angular/common": "2.0.0-rc.4",
+    "@angular/compiler": "2.0.0-rc.4",
+    "@angular/core": "2.0.0-rc.4",
+    "@angular/platform-browser": "2.0.0-rc.4",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.4",
+    "@angular/platform-server": "2.0.0-rc.4",
+    "@angular/router": "3.0.0-beta.2",
+    "email-validator": "1.0.4",
+    "everlive-sdk": "1.7.1",
+    "nativescript-angular": "0.3.0",
+    "nativescript-iqkeyboardmanager": "1.0.1",
+    "nativescript-social-share": "1.3.0",
+    "nativescript-unit-test-runner": "0.3.3",
+    "tns-core-modules": "2.2.0"
+  },
+  "devDependencies": {
+    "babel-traverse": "6.8.0",
+    "babel-types": "6.8.1",
+    "babylon": "6.8.0",
+    "codelyzer": "0.0.28",
+    "filewalker": "0.1.2",
+    "jasmine-core": "^2.4.1",
+    "karma": "^0.13.22",
+    "karma-jasmine": "^1.0.2",
+    "karma-nativescript-launcher": "^0.4.0",
+    "lazy": "1.0.11",
+    "nativescript-dev-typescript": "^0.3.2",
+    "tslint": "^3.14.0",
+    "typescript": "^1.8.10"
+  }
 }
